feat: implement container resolution stack and locking mechanisms
