+++ src/diwire/container_context.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
     from diwire.container import Container, ScopedContainer
 
 # Import signature builder to exclude Injected parameters from signature
-from diwire.container import _build_signature_without_injected
+from diwire.container_helpers import _build_signature_without_injected
 
 T = TypeVar("T")
 _C = TypeVar("_C", bound=type)
